quic: increase length of quicsocketconfig_buffer
Currently the length of this buffer is set to the enum index which is
zero based. This will cause an error when running in debug mode as there
is a check in AliasedBuffer::GetValue function that the index be less
than the length, but in this case they will be the same (11).

This commit increased the length by one.

PR-URL: https://github.com/nodejs/quic/pull/13
Reviewed-By: James M Snell <jasnell@gmail.com>